refactor(pkg/trie): Create TrieLookup structure
